pyproject.toml
[tool.poetry]
name = "drfprojecttemplate"
version = "0.1.0"
description = ""
authors = ["Jaime Alvarez <j.alvarez.gomez@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
requests = "^2.28.2"
pyyaml = "^6.0"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.3"
drf-spectacular = "^0.25.1"
django-sequences = "^2.7"
django-filter = "^22.1"
django-schema-graph = "^2.2.1"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
